Safari-609.1.11.1/LayoutTests/css3/flexbox/flexitem-no-margin-collapsing.html

<!DOCTYPE html>
<html>
<head>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox {
    background-color: lightgrey;
}
.flexbox p {
    height: 100px;
    width: 100px;
    margin: 10px;
    background-color: blue;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
</head>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="flexbox" style="position: relative">
<div data-offset-x="0" data-offset-y="0" data-expected-width="120" data-expected-height="120">
    <p data-offset-x="10" data-offset-y="10" data-expected-width="100" data-expected-height="100"></p>
</div>
</div>

</body>
</html>
